+++ app/Controllers/LoginController.php
@@ -3,8 +3,6 @@
 namespace App\Controllers;
 
 
-use App\Database\DB;
-use App\Web\Session;
 use Slim\Http\Request;
 use Slim\Http\Response;
 
@@ -18,7 +16,7 @@ class LoginController extends Controller
 	 */
 	public function show(Request $request, Response $response): Response
 	{
-		if (Session::get('logged', false)) {
+		if ($this->session->get('logged', false)) {
 			return redirect($response, 'home');
 		}
 		return $this->view->render($response, 'auth/login.twig');
@@ -32,29 +30,29 @@ class LoginController extends Controller
 	public function login(Request $request, Response $response): Response
 	{
 
-		$result = DB::query('SELECT `id`, `email`, `username`, `password`,`is_admin`, `active` FROM `users` WHERE `username` = ? OR `email` = ? LIMIT 1', [$request->getParam('username'), $request->getParam('username')])->fetch();
+		$result = $this->database->query('SELECT `id`, `email`, `username`, `password`,`is_admin`, `active` FROM `users` WHERE `username` = ? OR `email` = ? LIMIT 1', [$request->getParam('username'), $request->getParam('username')])->fetch();
 
 		if (!$result || !password_verify($request->getParam('password'), $result->password)) {
-			Session::alert(lang('bad_login'), 'danger');
+			$this->session->alert(lang('bad_login'), 'danger');
 			return redirect($response, 'login');
 		}
 
 		if (!$result->active) {
-			Session::alert(lang('account_disabled'), 'danger');
+			$this->session->alert(lang('account_disabled'), 'danger');
 			return redirect($response, 'login');
 		}
 
-		Session::set('logged', true);
-		Session::set('user_id', $result->id);
-		Session::set('username', $result->username);
-		Session::set('admin', $result->is_admin);
-		Session::set('used_space', humanFileSize($this->getUsedSpaceByUser($result->id)));
+		$this->session->set('logged', true);
+		$this->session->set('user_id', $result->id);
+		$this->session->set('username', $result->username);
+		$this->session->set('admin', $result->is_admin);
+		$this->session->set('used_space', humanFileSize($this->getUsedSpaceByUser($result->id)));
 
-		Session::alert(lang('welcome', [$result->username]), 'info');
+		$this->session->alert(lang('welcome', [$result->username]), 'info');
 		$this->logger->info("User $result->username logged in.");
 
-		if (Session::has('redirectTo')) {
-			return $response->withRedirect(Session::get('redirectTo'));
+		if ($this->session->has('redirectTo')) {
+			return $response->withRedirect($this->session->get('redirectTo'));
 		}
 
 		return redirect($response, 'home');
@@ -67,9 +65,9 @@ class LoginController extends Controller
 	 */
 	public function logout(Request $request, Response $response): Response
 	{
-		Session::clear();
-		Session::set('logged', false);
-		Session::alert(lang('goodbye'), 'warning');
+		$this->session->clear();
+		$this->session->set('logged', false);
+		$this->session->alert(lang('goodbye'), 'warning');
 		return redirect($response, 'login.show');
 	}