+++ app/Middleware/AdminMiddleware.php
@@ -3,20 +3,11 @@
 namespace App\Middleware;
 
 use App\Exceptions\UnauthorizedException;
-use App\Web\Session;
 use Slim\Http\Request;
 use Slim\Http\Response;
 
-class AdminMiddleware
+class AdminMiddleware extends Middleware
 {
-	/** @var \Slim\Container */
-	private $container;
-
-	public function __construct($container)
-	{
-		$this->container = $container;
-	}
-
 	/**
 	 * @param Request $request
 	 * @param Response $response
@@ -26,10 +17,10 @@ class AdminMiddleware
 	 */
 	public function __invoke(Request $request, Response $response, callable $next)
 	{
-		if (!$this->container->database->query('SELECT `id`, `is_admin` FROM `users` WHERE `id` = ? LIMIT 1', [Session::get('user_id')])->fetch()->is_admin) {
-			Session::alert('Your account is not admin anymore.', 'danger');
-			Session::set('admin', false);
-			Session::set('redirectTo', (isset($_SERVER['HTTPS']) ? "https" : "http") . "://$_SERVER[HTTP_HOST]$_SERVER[REQUEST_URI]");
+		if (!$this->database->query('SELECT `id`, `is_admin` FROM `users` WHERE `id` = ? LIMIT 1', [$this->session->get('user_id')])->fetch()->is_admin) {
+			$this->session->alert('Your account is not admin anymore.', 'danger');
+			$this->session->set('admin', false);
+			$this->session->set('redirectTo', (isset($_SERVER['HTTPS']) ? "https" : "http") . "://$_SERVER[HTTP_HOST]$_SERVER[REQUEST_URI]");
 			throw new UnauthorizedException();
 		}