+++ resources/lang/it.lang.php
@@ -85,4 +85,5 @@ return [
 	'cannot_delete' => 'Non puoi eliminare te stesso.',
 	'cannot_demote' => 'Non puoi degradare te stesso. ',
 	'cannot_write_file' => 'Il percorso di destinazione non è scrivibile.',
+	'deleted_orphans' => 'Eliminati %d file orfani.',
 ];